+++ modules/mtl/arrow-mtl-data/src/test/kotlin/arrow/mtl/KleisliTest.kt
@@ -1,4 +1,4 @@
-package arrow.data
+package arrow.mtl
 
 import arrow.Kind
 import arrow.core.Either
@@ -15,10 +15,9 @@ import arrow.core.extensions.`try`.monadError.monadError
 import arrow.core.extensions.const.divisible.divisible
 import arrow.core.extensions.id.monad.monad
 import arrow.core.extensions.monoid
-import arrow.core.value
-import arrow.data.extensions.kleisli.contravariant.contravariant
-import arrow.data.extensions.kleisli.divisible.divisible
-import arrow.data.extensions.kleisli.monadError.monadError
+import arrow.mtl.extensions.kleisli.contravariant.contravariant
+import arrow.mtl.extensions.kleisli.divisible.divisible
+import arrow.mtl.extensions.kleisli.monadError.monadError
 import arrow.test.UnitSpec
 import arrow.test.laws.BracketLaws
 import arrow.test.laws.ContravariantLaws
@@ -32,8 +31,8 @@ import arrow.typeclasses.conest
 import arrow.typeclasses.const
 import arrow.typeclasses.counnest
 import arrow.typeclasses.value
-import io.kotlintest.shouldBe
 import io.kotlintest.runner.junit4.KotlinTestRunner
+import io.kotlintest.shouldBe
 import org.junit.runner.RunWith
 
 @RunWith(KotlinTestRunner::class)
@@ -75,9 +74,9 @@ class KleisliTest : UnitSpec() {
     "andThen should continue sequence" {
       val kleisli: Kleisli<ForId, Int, Int> = Kleisli { a: Int -> Id(a) }
 
-      kleisli.andThen(Id.monad(), Id(3)).run(0).value() shouldBe 3
+      kleisli.andThen(Id.monad(), Id(3)).run(0) shouldBe 3
 
-      kleisli.andThen(Id.monad()) { b -> Id(b + 1) }.run(0).value() shouldBe 1
+      kleisli.andThen(Id.monad()) { b -> Id(b + 1) }.run(0) shouldBe 1
     }
   }
 }
