+++ modules/optics/arrow-optics/src/main/kotlin/arrow/optics/Lens.kt
@@ -8,8 +8,9 @@ import arrow.core.Some
 import arrow.core.Tuple2
 import arrow.core.identity
 import arrow.core.toT
-import arrow.data.Reader
+import arrow.mtl.Reader
 import arrow.data.State
+import arrow.mtl.map
 import arrow.data.map
 import arrow.higherkind
 import arrow.typeclasses.Applicative
